Time: 0 ms (100%), Space: 7.9 MB (11.05%) - LeetHub

+++ 1812-determine-color-of-a-chessboard-square/1812-determine-color-of-a-chessboard-square.cpp
@@ -1,6 +1,6 @@
 class Solution {
 public:
     bool squareIsWhite(string coordinates) {
-       return ((int)coordinates[0] +  coordinates[1] ) %2==1;
+       return (coordinates[0] +  coordinates[1] ) %2==1;
      }
 };
